% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assign_rownames}
\alias{assign_rownames}
\title{Make first column of matrix the rownames and drop}
\usage{
assign_rownames(embedding)
}
\arguments{
\item{embedding}{A dataframe.}
}
\value{
The same dataframe with the first row as the row.names
}
\description{
Make first column of matrix the rownames and drop
}
\examples{
embedding <- assign_rownames(embedding)
}
