% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dropMT}
\alias{dropMT}
\title{Drop Empty}
\usage{
dropMT(x)
}
\arguments{
\item{x}{A dataframe, matrix, or vector.}
}
\value{
The same object without any empty rows
}
\description{
Drop empty rows from a dataframe
}
