% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exponent}
\alias{exponent}
\title{Exponentiate}
\usage{
exponent(x, power)
}
\arguments{
\item{x}{The number or vector to exponentiate}

\item{power}{The power to exponentiate to}
}
\description{
Exponentiate a number or vector to a given power
}
