% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{howmuch}
\alias{howmuch}
\title{Count the number of unique elements in a list or vector}
\usage{
howmuch(x)
}
\arguments{
\item{x}{A list or vector}
}
\value{
the number of unique elements in the vector or list
}
\description{
Count the number of unique elements in a list or vector
}
\examples{
x 
}
