% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_dim}
\alias{make_dim}
\title{Make semantic dimension}
\usage{
make_dim(embedding, pairs)
}
\arguments{
\item{embedding}{A word embedding in the form of a matrix}

\item{pairs}{A set of antonym pairs}
}
\value{
A dataframe with the projections
}
\description{
Make semantic dimension
}
