% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{suppress_ticks}
\alias{suppress_ticks}
\title{Suppress tick marks}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 4.
}
\usage{
suppress_ticks
}
\description{
For suppressing tick labels in ggplot
}
\examples{
ggplot(df, aes(x = VAR1, y = VAR2)) + geom_scatter() + suppress_ticks
}
\keyword{datasets}
