% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{twenty3pal}
\alias{twenty3pal}
\title{2023 Color Palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
twenty3pal
}
\description{
A color palette for the year 2023.
}
\keyword{datasets}
